refactor: rename beefy classes according to name convention
